+++ CHANGELOG.md
@@ -19,6 +19,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - #2824, Fix range request with 0 rows and 0 offset return status 416 - @strengthless
  - #3015, Fix unnecessary count() on RPC returning single - @steve-chavez
 
+### Fixed
+
+ - #2824, Fix regression by reverting fix that returned 206 when first position = length in a `Range` header - @laurenceisla, @strengthless
+
 ## [11.2.1] - 2023-10-03
 
 ### Fixed